requirements.txt
requests
beautifulsoup4
pandas
pyyaml
matplotlib
seaborn